requirements.txt
fastapi==0.101.0
pydantic==2.1.1
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.3
python-multipart==0.0.6